requirements.txt
django>=4.2.0,<5.0.0
python-decouple>=3.8
Pillow>=10.0.0
djangorestframework>=3.14.0
djangorestframework-api-key>=3.0.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
django-filter>=23.0.0